% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_solver.R
\name{by_deduction}
\alias{by_deduction}
\title{By deduction function}
\usage{
by_deduction(emptycells)
}
\arguments{
\item{emptycells}{The dataframe with information on the empty cells
of the matrix.}
}
\value{
a 9x9 Matrix corresponding to the solution found by the solver,
or TRUE/FALSE
}
\description{
This function is part of the matrix_solver function
to solve Sudoku grids. For each possibility, the function checks if there is
another case in the same line/col/box with the same possibility.
If not, the number corresponding to the possibility should be the solution.
}
\author{
Flavie B.
}
