% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_solver.R
\name{empty_cells}
\alias{empty_cells}
\title{Empty cells function
This function finds empty cells in the sudoku grid and searches for
possible numbers to fill them.}
\usage{
empty_cells(Matrix_m)
}
\arguments{
\item{Matrix_m}{The playable matrix generated.}
}
\value{
a list containing a dataframe with information on the empty cells
of the matrix and a Boolean
}
\description{
Empty cells function
This function finds empty cells in the sudoku grid and searches for
possible numbers to fill them.
}
\author{
Flavie B.
}
