% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_grid.R
\name{generate3}
\alias{generate3}
\title{Generate 3 function}
\usage{
generate3(Temp_m)
}
\arguments{
\item{Temp_m}{A temporary matrix storing the three columns of one block
from the main matrix}
}
\value{
The same temporary matrix with the rest of the columns filled
(doing another circle shift)
}
\description{
Generate 3 function
}
\examples{
Matrix_m <- matrix(sample(1:9, replace = FALSE), nrow = 9, ncol = 9)
Matrix_m[,1:3] <- generate3(Matrix_m[,1:3])
}
\author{
Flavie B.
}
