% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_solver.R
\name{is_valid}
\alias{is_valid}
\title{Is valid function}
\usage{
is_valid(matrix, num, row, col)
}
\arguments{
\item{matrix}{The playable matrix generated.}

\item{num}{the number to try (an integer between 1 and 9)}

\item{row}{the row index}

\item{col}{the column index}
}
\value{
False if the number already appears in the row/column/block,
True otherwise.
}
\description{
This function checks if a certain number can be placed in a
certain empty case. The number is valid if it does not already appear
in the row, column, or block of the case.
}
\author{
Flavie B.
}
