% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch App function}
\usage{
launch_app()
}
\value{
shiny app
}
\description{
Launch the Shiny app "sudoku_grids". The directory "sudoku_grids" needs to be set as the working directory.
}
\author{
Flavie B.
}
