% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_solver.R
\name{matrix_solver}
\alias{matrix_solver}
\title{Matrix solver function
This function solves a Sudoku grid.}
\usage{
matrix_solver(Matrix, layer = 0)
}
\arguments{
\item{Matrix}{The playable matrix generated.}

\item{layer}{Integer showing depth of the called functions. Default = 0.}
}
\value{
a 9x9 Matrix corresponding to the solution found by the solver,
or TRUE/FALSE
}
\description{
Matrix solver function
This function solves a Sudoku grid.
}
\author{
Flavie B.
}
