% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_grid.R
\name{mix3}
\alias{mix3}
\title{Mix 3 function}
\usage{
mix3(Matrix_m)
}
\arguments{
\item{Matrix_m}{A complete matrix generated}
}
\value{
The same matrix with the order of the lines and columns shuffled
inside a group of 3
}
\description{
Mix 3 function
}
\author{
Flavie B.
}
