% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_grid.R
\name{offset3}
\alias{offset3}
\title{Offset 3 function}
\usage{
offset3(Vect, Rev = FALSE)
}
\arguments{
\item{Vect}{A vector of 9 corresponding to the line or column of interest}

\item{Rev}{A Boolean to indicate the direction of the offset}
}
\value{
The same vector with the first three numbers now being the fourth fifth and sixth and so on (shifting in circle)
}
\description{
Offset 3 function
}
\examples{
Matrix_m <- matrix(sample(1:9, replace = FALSE), nrow = 9, ncol = 9)
Matrix_m[2,] <- offset3(Matrix_m[1,])
}
\author{
Flavie B.
}
