% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_solver.R
\name{trials_errors}
\alias{trials_errors}
\title{Trials and errors function}
\usage{
trials_errors(matrix_m, emptycells, layer)
}
\arguments{
\item{matrix_m}{The generated matrix/Sudoku grid we are trying to solve.}

\item{emptycells}{The dataframe with information on the empty cells
of the matrix.}

\item{layer}{Integer showing depth of the called functions.}
}
\value{
a 9x9 Matrix corresponding to the solution found by the solver,
or TRUE if the Sudoku grid has multiple solutions
(and FALSE if something went wrong).
}
\description{
This function is part of the matrix_solver function
to solve Sudoku grids.
It uses a "trial and error" method to solve the Sudoku grid.
The function tries each possibility from emptycells until the grid is solved.
It is also used to check if the Sudoku grid generated has a unique solution
(and therefore is playable) or has multiple solutions.
}
\author{
Flavie B.
}
