% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_grid.R
\name{matrix_grid}
\alias{matrix_grid}
\title{Matrix grid function
Generates a filled 9x9 matrix we will be able to use as a sudoku grid}
\usage{
matrix_grid()
}
\value{
a 9x9 Matrix
}
\description{
Matrix grid function
Generates a filled 9x9 matrix we will be able to use as a sudoku grid
}
\author{
Flavie B.
}
