% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_remover.R
\name{matrix_remover}
\alias{matrix_remover}
\title{Matrix remover function
A sudoku grid ready to be played based on the solution matrix generated.
The function clears some of the cases (replacing the numbers with NA)}
\usage{
matrix_remover(Matrix_m)
}
\arguments{
\item{Matrix_m}{The solution matrix generated}
}
\value{
a 9x9 Matrix
}
\description{
Matrix remover function
A sudoku grid ready to be played based on the solution matrix generated.
The function clears some of the cases (replacing the numbers with NA)
}
\author{
Flavie B.
}
